test-requirements.txt
nose
PyHamcrest
pep8
mock
coverage
